Изменения для улучшения читаемости кода.

+++ lab1/Регистр сдвига/BitArray.h
@@ -6,14 +6,14 @@
 
 typedef struct {
     uint8_t *bucket;
-    long long unsigned length;
+    uint64_t length;
 } BitArray;
 
-int initBitArray(BitArray *array, long long unsigned length);
-long long unsigned getBitArrayLength(BitArray *array);
-uint8_t getBitArrayElement(BitArray *array, long long unsigned i);
-void setBitArrayElement(BitArray *array, long long unsigned i, uint8_t element);
+int initBitArray(BitArray *array, uint64_t length);
+uint64_t getBitArrayLength(BitArray *array);
+uint8_t getBitArrayElement(BitArray *array, uint64_t i);
+void setBitArrayElement(BitArray *array, uint64_t i, uint8_t element);
 void freeBitArray(BitArray *array);
-int readArrayFromFile(BitArray *array, long long unsigned length, FILE *fp);
+int readArrayFromFile(BitArray *array, uint64_t length, FILE *fp);
 
 #endif
